// project-rekord/src/app/enums/chestChanceTypes.ts
export enum ChestChanceTypes {
    MOVE = "move",
    MOVE_NEAREST = "movenearest",
    ADD_FUNDS = "addfunds",
    ADD_FUNDS_FROM_PLAYERS = "addfundsfromplayers",
    REMOVE_FUNDS = "removefunds",
    REMOVE_FUNDS_TO_PLAYERS = "removefundstoplayers",
    JAIL = "jail",
    PROPERTY_CHARGES = "propertycharges",
}